feat: 음수를 전달할 경우 RuntimeException 예외 구현
