+++ src/main/java/calculator/MessageConstant.java
@@ -1,0 +1,6 @@
+package calculator;
+
+public class MessageConstant {
+    public static final String NEGATIVE_NUMBER_EXCEPTION = "음수는 포함될 수 없습니다.";
+
+}
