+++ src/test/java/calculator/StringAdderCalculatorTest.java
@@ -1,10 +1,14 @@
+package calculator;
+
+import calculator.Calculator;
+import calculator.Prompt;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 public class StringAdderCalculatorTest {
 
@@ -37,8 +41,16 @@ public class StringAdderCalculatorTest {
     @Test
     @DisplayName("“//”와 “\\n” 문자 사이에 커스텀 구분자를 지정할 수 있다.")
     public void addIntegerSeparatedByCustomDelimiter() {
-
         assertThat(calculator.calculate(new Prompt("//;\n1;2;3"))).isEqualTo(6);
+    }
+
+    @Test
+    @DisplayName("음수를 전달할 경우 RuntimeException 예외가 발생해야 한다.")
+    public void isExistNegativeNumber() {
+        assertThatExceptionOfType(RuntimeException.class)
+                .isThrownBy(() -> {
+                    calculator.calculate(new Prompt("-1:2:3"));
+                });
     }
 }
